+++ models/users.js
@@ -3,13 +3,13 @@ var db = require('../db');
 
 module.exports = db.define('users', {
     id: {type: types.INTEGER, autoIncrement: true, unique: true, primaryKey: true},
-    status: types.ENUM('new', 'active', 'inactive'),
+    status: types.ENUM('new', 'active', 'inactive', 'closed'),
     username: types.STRING(20),
     first: types.STRING(40),
     last: types.STRING(40),
     email: types.STRING(60),
     phone: types.STRING(20),
-    gender: types.ENUM('M', 'F', '?'),
+    gender: types.ENUM('M', 'F', 'O', 'P'),
     occupation: types.STRING(40),
     company: types.STRING(40),
     website: types.STRING(80),
@@ -28,13 +28,22 @@ module.exports = db.define('users', {
             return this.create(model);
         },
         save: function (id, model) {
-            return this.update(model, {where: {id: id}});
+            if (this.SESSION_USER === false) {
+                return 401;
+            }
+            return this.update(model, {where: {id: id, user_id: this.SESSION_USER}});
         },
         remove: function (id) {
-            return this.destroy({where: {id: id}});
+            if (this.SESSION_USER === false) {
+                return 401;
+            }
+            return this.destroy({where: {id: id, user_id: this.SESSION_USER}});
         },
         queryize: function (params) {
             return {};
+        },
+        getID: function (username) {
+            return this.findOne({where: types.or({username: username}, {email: username}) });
         }
     }
 });
